Code refactoring

+++ src/main/java/com/example/MyBookShopApp/model/entities/Book/BookReviewLikeRepository.java
@@ -1,5 +1,8 @@
 package com.example.MyBookShopApp.model.entities.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookReviewLikeRepository extends JpaRepository<BookReviewLike, Long> {
+public interface BookReviewLikeRepository extends JpaRepository<BookReviewLike, Integer> {
+
+    
+
 }
